Updated zstd go library to avoid cgo dependency for compression
