--- pkg/compression/zstd_native.go
@@ -1,64 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-// +build !cgo
-
-// Pure GO ZStd library only supports decompression
-
-package compression
-
-import (
-	"bytes"
-	"github.com/klauspost/compress/zstd"
-	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
-)
-
-// Pure GO ZStd library only supports decompression
-
-type zstdProvider struct {
-}
-
-func NewZStdProvider() Provider {
-	return &zstdProvider{}
-}
-
-func (zstdProvider) CanCompress() bool {
-	return false
-}
-
-func (zstdProvider) Compress(data []byte) []byte {
-	log.Panic("ZSTD compression is not supported if CGo is disabled")
-	return nil
-}
-
-func (zstdProvider) Decompress(compressedData []byte, originalSize int) ([]byte, error) {
-	d, err := zstd.NewReader(bytes.NewReader(compressedData))
-	if err != nil {
-		return nil, err
-	}
-
-	uncompressed := make([]byte, originalSize)
-	size, err := d.Read(uncompressed)
-	if err != nil {
-		return nil, err
-	} else if size != originalSize {
-		return nil, errors.New("Invalid uncompressed size")
-	} else {
-		return uncompressed, nil
-	}
-}
